Switch from a ScrollView to a proper ListView.
This way we're only rendering the Message tiles that are actually on the
screen at any given moment, and the rendering is therefore deferred and
will not lag the app / use memory as much.

This eliminates the custom PointyScrollView class and moves the
renderStreamMessage code to a separate class which is a custom ArrayAdapter
for our ListView.