Update app for changes in zev-api-rework-2.
This simplified message recieving so that "everything is a huddle",
which makes the code look a lot cleaner.

There are some things added for compatibility with the existing code on
production that will not be required or useful once Zev's work is pushed
fully; at this point this compat code can be removed.