Add Ant build files
This allows building with Ant as well as with Eclipse (both build
methods can coexist).

The local.properties file isn't committed. To generate one for your
system, make sure the tools/ directory of the SDK is in your path, and
run:

    android update project --target TARGETNUM --path .

where TARGETNUM is the target in `android list targets` that
corresponds to API version 17, our current target.