Add support for narrowing by stream.
No old messages are fetched from the server. When new messages are received
they are checked against a pattern to see if they apply to the filter and
if so they are added to the narrowed view.

This code is not compatible with Android API version 10, and is not used
on those platforms, I hope.