[DCU] Use string list for recipients rather than join table.
ORMLite does a really bad job with the join table. We never look up
"messages to person X", even on the web app. With this, each PM or
huddle is just a compare on the simple string. An index on that field is
nearly equivalent to the recipient + huddle tables used on the server.

Suboptimal from a database design perspective, sure. But a lot simpler.