Log or die on exceptions rather than just printing and continuing.
Now that we have Crashlytics, its reasonable for us to log the
exceptions we were previously just leaving on the floor.

Sometimes, an exception is obviously fatal and indicates a horrible
error condition, like a SQLException. These are RuntimeExceptions. Other
times, the exception might be ignorable, in which case we just log and
continue.