Fix duplicated fragment on repeated narrows.
The FragmentTransaction API is a little confusing. When the addToBackStack
docs say "This means that the transaction will be remembered after it is
committed, and will reverse its operation when later popped off the
stack.", they really  mean it. Only that transaction is un-done on back,
not any subsequent transactions that aren't added to the back stack.

So, when you narrow twice and then hit back, it un-did the
home->first narrow transaction, ignores that the first narrow's fragment
is no longer there, and left the second narrow's fragment alone.