Better notification display options
This uses the default layout instead of the BigText layout. That
makes the text show properly when more than one notification is active,
fixes compatibility with older Android versions, and shows the
Zulip icon in the notification tile.

However, this limits the message preview to one line. In Android >= 4.1,
there is a secondary line available, which is used to display information
on the narrow.