Don't show search pop up for honeycomb++ search (appears in actionbar instead)
For apps that are Honeycomb+, clicking search was showing 2 search fields: a pop up dialog and a field in the Action Bar.  Search in the ActionBar only happens for Honeycomb+; the popup dialog should only happen for < Honeycomb. This changes the logic to match that: the search dialog will only pop up now when there is no search in the ActionBar.