Add DrawCustomView.java for drawing on canvas.
If same file path is used, the image is cached based on file path by
Glide in PhotoEditActivity. So if the user changes the image in
SendActivity (after pressing back from EditActivity), the changes are not
reflected back in EditActivity.

After glide loads a bitmap in an imageview, it uses its own drawable
GlideBitmapDrawable to place inside the imageView unless we explicitly
use .asBitmap() option. Hence, instanceOf should to used to distinguish
between the two types.