Add photo uploads, cropping and marking tools.
* Set up camera button.

Launch MediaStore.ACTION_IMAGE_CAPTURE intent on clicking camera button.

* Set up file provider to get full sized image.

The path for images directory in the app will be different for debug and
realese builds. In order to use getExternalStorageDir() we need to ask
write permission for android 4.3 and below.

* Disable camera button while sending message.

* Add PhotoSendActivity to crop captured photo.

Android-image-cropper library is used for cropping functionality.

* Use Glide library to decode bitmaps with high performance.

* Set up delete, crop and send button in PhotoSendActivity.

* Set up edit button in PhotoSendActivity to launch PhotoEditAcitvity.

* Set up layout for PhotoEditActivity.

* Add DrawCustomView.java for drawing on canvas.

If same file path is used, the image is cached based on file path by
Glide in PhotoEditActivity. So if the user changes the image in
SendActivity (after pressing back from EditActivity), the changes are not
reflected back in EditActivity.

After glide loads a bitmap in an imageview, it uses its own drawable
GlideBitmapDrawable to place inside the imageView unless we explicitly
use .asBitmap() option. Hence, instanceOf should to used to distinguish
between the two types.

* Correct file paths in PhotoHelper.saveBitmapAsFile().

* Add colors and undo, back, send and crop button.

This adds more color options to marker tool and sets up undo, back, send
and crop buttons in PhotoEditActivity.

* Load image when PhotoSendActivity is in focus.

* Add cancel button and override back button function.

These changes are committed in PhotoSendActivity. The cancel button
takes user back to ZulipActivity.

In case user has cropped the image, on pressing back these changes should
be undone otherwise user should be sent back to ZulipActivity.

* Activate crop button after redirection from PhotoEditActivity.

During crop, after user is redirected to PhotoSendActivity from
PhotoEditActivity, the crop button should appear in an active state.

Fixes #228.