Maintain scroll position in case of Load position ABOVE messages.
Account for message headers generated when messages are loaded above a
list of messages.

This is important because message headers are also counted as items in the
recyclerview and hence when retaining the position of visible message at
the top of list, we need to account for all the message headers generated
for the set of LoadPosition.ABOVE messages.

Fixes #353.